Completed Lobby (not tested with BE yet)
