+++ src/components/views/Lobby.js
@@ -1,236 +1,342 @@
-import {
-    Button,
-    FormControl,
-    Grid,
-    InputLabel,
-    Select,
-    TextField,
-    Typography,
-} from "@material-ui/core";
-import { useHistory } from "react-router-dom";
-import { useStyles } from './styles/lobbyStyles';
-import {useState} from "react";
+import {AppBar, Button, Card, CardActions, CardContent, Grid, TextField, Toolbar, Typography} from "@material-ui/core";
+import {useEffect, useState} from "react";
+import {getDomain} from "../../helpers/getDomain";
+import Stomp from "stompjs";
+import {useLocation, useParams} from "react-router-dom";
+import SockJS from "sockjs-client";
+import {api} from "../../helpers/api";
+import PropTypes from "prop-types";
+import {makeStyles} from "@material-ui/core/styles";
+import IconButton from "@material-ui/core/IconButton";
+import ExitToAppIcon from '@material-ui/icons/ExitToApp';
+import GameSettings from "../ui/GameSettings";
+import Player from "../../models/Player";
+import SettingsData from "../../models/SettingsData";
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+        height: '100vh',
+        direction: 'column',
+        alignItems: 'center',
+        justifyContent: 'center'
+    },
+    mainCard: {
+        width: '80vw',
+        height: '80vh',
+    },
+    cardBar: {
+        alignContent: 'center',
+        justifyContent: 'space-between',
+        height: '20%'
+    },
+    lobbyTitle: {
+        variant: 'h4',
+        color: 'inherit'
+    },
+    copyButton: {
+        variant: "outlined",
+        color: "secondary"
+    },
+    cardContainer: {
+        // this probably needs changing
+        height: '80%',
+        width: '100%',
+        direction: "row",
+        alignItems: 'center',
+        justifyContent:'space-around'
+    },
+    // may need to move playerlist component into view component to access classes
+    playerListContainer: {
+        width: "100%",
+        height: "100%",
+        direction: "column",
+        justifyContent: "center",
+        alignItems: "center",
+        spacing: 2
+    },
+    playerCard: {
+
+    },
+    settings: {
+
+    }
+}));
+const PlayerList = (props) => {
+
+    return (
+        <Grid container
+              width={"30%"}
+              height={"100%"}
+              direction={"column"}
+              justifyContent={"center"}
+              alignItems={"center"}
+              spacing={2}
+        >
+            {props.list.map((player, index) => {
+                return (
+                    <Grid item key={index}>
+                        <Card>
+                            <CardContent>
+                                <Typography variant={"h6"}>
+                                    {player.username}
+                                </Typography>
+                            </CardContent>
+                        </Card>
+                    </Grid>
+                )
+            })}
+        </Grid>
+    )
+}
+
+PlayerList.propTypes = {
+    list: PropTypes.array
+}
 
 const Lobby = () => {
+    // maybe start using contexts to pass state
+    // TODO: disconnect from WS endpoint on unmount and send leave game message to WS endpoint
+    // TODO: add leave game button
 
     const classes = useStyles();
-    const history = useHistory();
+    const location = useLocation();
+    const { gameId } = useParams();
+
+    const [gameStarting, setGameStarting] = useState(false);
+    const [user, setUser] = useState(location.state.user);
+    const [players, setPlayers] = useState([]);
+
+    // setting state variables
+    const [isHost, setIsHost] = useState(false);
+
     const [language, setLanguage] = useState("en");
-    const [playlist, setPlaylist] = useState("");
-    const [numPlayers, setNumPlayers] = useState(3);
-    const [showStartInfo, setShowStartInfo] = useState(true);
+    const [playlistUrl, setPlaylistUrl] = useState("");
     const [balance, setBalance] = useState(3000);
     const [bigBlind, setBigBlind] = useState(100);
     const [smallBlind, setSmallBlind] = useState(50);
 
-    //this might complicate things and is probably unnecessary for m3
-    const handleUserClick = (id) => {
-        // redirect to user profile (IF user is registered)
-        // somehow maintain WS connection though?
-        // maybe make profile info a dialog instead of a new page?
-    }
 
-    const handleBackClick = () => {
-        // leave game
-        // disconnect WS
-        // redirect to home
-    }
+    const [stompClient, setStompClient] = useState(null);
+    // subscribe to player-list updates
+    const [playersSubscription, setPlayersSubscription] = useState(null);
+    // subscribe to setting updates
+    const [settingsSubscription, setSettingsSubscription] = useState(null);
+    // subscribe to game started
+    const [gameStartSubscription, setGameStartSubscription] = useState(null);
 
-    const handleStartGame = () => {
-        const data = {
-            language,
-            playlist,
-            numPlayers,
-            showStartInfo,
-            balance,
-            bigBlind,
-            smallBlind,
-        };
+    // when component first mounts:
+    // 1. create a new websocket connection
+    // 2. subscribe to the /topic/games/{gameId}/players topic
+    // 3. send a message to the /app/games/{gameId}/players app endpoint
+    // subscription must be done first, to update the players list with own name
+    useEffect(() => {
+        // check if user is host -> able to modify settings
+        const checkUserHost = async () => {
+            const response = await api.get(`/games/${gameId}/host`);
+            const host = new Player(response.data);
+            if (host.token === user.token) {
+                setIsHost(true);
+            }
+            else {
+                setIsHost(false);
+            }
+        }
+        checkUserHost();
 
-        // Send the data to the backend using an API call
-        fetch("/api/startGame", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json",
-            },
-            body: JSON.stringify(data),
-        })
-            .then((response) => response.json())
-            .then((data) => {
-                console.log(data);
+        // connect WS
+        async function connectSocket() {
+            let socket = new SockJS("http://localhost:8080/sopra-websocket");
+            const client = Stomp.over(socket);
+            setStompClient(client);
 
-                // Navigate to the game page with the created game ID
-                history.push(`/games/${data.gameId}`);
-            })
-            .catch((error) => {
-                console.error("Error:", error);
+            client.connect({}, () => {
+                console.log("Websocket connection established");
+                // SUBSCRIPTIONS //
+                setPlayersSubscription(client.subscribe(
+                    `/topic/games/${gameId}/players`,
+                    (message) => {
+                        setPlayers(JSON.parse(message.data));
+                        console.log(message.data);
+                    }
+                ));
+                setSettingsSubscription(
+                    client.subscribe(
+                    `/topic/games/${gameId}/settings`,
+                    (message) => {
+                        console.log(message.data)
+                        const settingsData = new SettingsData(message.data)
+                        setLanguage(settingsData.language)
+                        setPlaylistUrl(settingsData.playlistUrl)
+                        setBalance(settingsData.balance)
+                        setBigBlind(settingsData.bigBlind)
+                        setSmallBlind(settingsData.smallBlind)
+                    }
+                ));
+                setGameStartSubscription( // -> if the game is started by someone else
+                    client.subscribe(
+                        `/topic/games/${gameId}/settings`,
+                        (message) => {
+                            console.log(message.data)
+                            setGameStarting(true) // -> so player is not removed from game on unmount
+                            history.push(`/games/${gameId}`)
+                        }
+                    )
+                )
+                // ADD PLAYER TO GAME
+                const username = user.username;
+                const score = 0;
+                const token = localStorage.getItem("token");
+                const requestBody = JSON.stringify({username, score, token});
+                try {
+                    client.send(
+                        `/app/games/${gameId}/players`,
+                        {},
+                        requestBody
+                    )
+                }
+                catch (error) {
+                    alert(`${error.response.message}`)
+                }
             });
-    };
+        }
+        connectSocket();
+
+        // CLEANUP //
+        return () => {
+            if(gameStarting){
+                gameStartSubscription.unsubscribe();
+                settingsSubscription.unsubscribe();
+                return
+            }
+            leaveGame()
+        }
+    }, []);
 
 
+    /** Dynamic Components */
+    let playerList = <PlayerList list={players}/>;
+    // rerender PlayerList if players change
+    useEffect(() => {
+        playerList = <PlayerList list={players}/>;
+    }, [players])
 
-    const lobby = (
-
-// define lobbyConTainer
-
-        <div className={classes.lobbyContainer}>
-
-            <div className={classes.titleContainer}>
-                <Typography variant="h3" className={classes.title}>
-                    TESTUSER's LOBBY
-                </Typography>
-
-                <div className={classes.lobbyCodeContainer}>
-                    <Typography className={classes.lobbyCode}>CODE: XYZ123</Typography>
-                    <Button variant="contained" color="primary">
-                        COPY
-                    </Button>
-                </div>
-
-                <svg
-                    className={classes.backIcon}
-                    viewBox="0 0 24 24"
-                    onClick={handleBackClick}
-                >
-                    <path d="M24 11.5H3.83l7.88-7.88L10.5.62.62 10.5l9.71 9.71 1.41-1.41-7.88-7.88H24v-2z" />
-                </svg>
-
-            </div>
-
-            <Grid container spacing={3}>
-            </Grid>
-
-            <Grid item xs={12} md={5}>
+    let settings = <GameSettings isHost={isHost}
+                                 // variables
+                                 language={language}
+                                 initialBalance={balance}
+                                 bigBlind={bigBlind}
+                                 smallBlind={smallBlind}
+                                 playlist={playlistUrl}
+                                 // setters
+                                 onLanguageChange={l => setLanguage(l)}
+                                 onInitialBalanceChange={b => setBalance(b)}
+                                 onBigBlindChange={b => setBigBlind(b)}
+                                 onSmallBlindChange={s => setSmallBlind(s)}
+                                 onPlaylistChange={p => setPlaylistUrl(p)}
+    />;
 
 
-                <div className={classes.lobbyMain}>
+    useEffect(() => {
+        settings = <GameSettings isHost={isHost}
+                                // variables
+                                 language={language}
+                                 balance={balance}
+                                 bigBlind={bigBlind}
+                                 smallBlind={smallBlind}
+                                 playlistUrl={playlistUrl}
+                                // setters
+                                 onLanguageChange={l => setLanguage(l)}
+                                 onBalanceChange={b => setBalance(b)}
+                                 onBigBlindChange={b => setBigBlind(b)}
+                                 onSmallBlindChange={s => setSmallBlind(s)}
+                                 onPlaylistUrlChange={p => setPlaylistUrl(p)}
+        />
+    }, [isHost, language, balance, bigBlind, smallBlind, playlistUrl])
 
+    const startGame = () => {
+        // prevent player being removed from game on unmount
+        setGameStarting(true)
 
-                    <div className={classes.playerListContainer}>
-                        <Typography variant="h6" className={classes.textPlayers}>
-                            PLAYERS
-                        </Typography>
+        // notify other players
+        try {
+            client.send(`/games/${gameId}/start`, {})
+        }
+        catch (error) {
+            const response = error.response
 
-                        <div className="lobby-players-list">
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('your-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>YOU</Typography>
-                            </div>
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('user2-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>TESTUSER2</Typography>
-                            </div>
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('user3-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>TESTUSER3</Typography>
-                            </div>
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('user4-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>TESTUSER4</Typography>
-                            </div>
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('user5-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>TESTUSER5</Typography>
-                            </div>
-                            <div className={classes.playerButtonContainer} onClick={() => handleUserClick('user6-id')}>
-                                <div className={classes.circle}></div>
-                                <Typography className={classes.playerName}>TESTUSER6</Typography>
-                            </div>
-                        </div>
+            if(response.status === 400) {
+                alert(`Couldn't start the game: ${response.message}`)
+            }
+            else {
+                console.log("Something went wrong while starting the game")
+            }
+        }
 
-                        <div className={classes.path}></div>
+        // redirect
+        history.push(`/games/${gameId}`)
+    }
 
+    const leaveGame = () => {
+        // remove player from playerList & disconnect WS
+        const destination = `/app/games/${gameId}/players/leave`
+        const token = localStorage.getItem("token");
+        const username = localStorage.getItem("username");
+        const responseBody = JSON.stringify({ token, username });
+        // TODO catch exceptions
+        stompClient.send(destination, {}, responseBody);
+        // unsub & disconnect
+        gameStartSubscription.unsubscribe();
+        settingsSubscription.unsubscribe();
+        stompClient.disconnect();
+    }
 
-                    </div>
-                </div>
-            </Grid>
-
-
-            <Grid item xs={12} md={5}>
-
-
-
-                <div className={classes.settingsContainer}>
-                    <Typography variant="h6" className={classes.textSettings}>SETTINGS</Typography>
-
-
-
-                    <div className={classes.settingsLeft}>
-
-                        <div className={classes.settingsTitle}>
-                            <Typography>MODIFY VIDEO-POOL</Typography>
-                        </div>
-
-                        <FormControl>
-                            <InputLabel htmlFor="lobby-language-select">Language</InputLabel>
-                            <Select
-                                className={classes.languageSelect}
-                                native
-                                inputProps={{
-                                    id: 'lobby-language-select'
-                                }}
+    return (
+        <Grid container
+              className={classes.root}
+        >
+            <Grid item>
+                <Card className={classes.mainCard}>
+                    <AppBar position={"relative"}>
+                        <Toolbar className={classes.cardBar}>
+                            <Typography className={classes.lobbyTitle}>
+                                {user.username}'s lobby
+                            </Typography>
+                            <Button variant={"contained"}
+                                    disabled={!isHost}
+                                    onClick={startGame}
                             >
-                                <option value="en">ENGLISH</option>
-                                <option value="de">GERMAN</option>
-                                <option value="zh">CHINESE</option>
-                            </Select>
-                        </FormControl>
-
-                        <div className={classes.settingsTitle}>
-                            <Typography>Playlist</Typography>
-                        </div>
-                        <TextField type="text" placeholder="Paste a link here" fullWidth />
-
-                        <div className={classes.settingsTitle}>
-                            <Typography>PLAYERS</Typography>
-                        </div>
-
-                        <div className="lobby-players-setting-container">
-                            <div className="lobby-players-setting-value">
-                                <TextField
-                                    className={classes.playersTextField}
-                                    placeholder="Enter number of players"
-                                    inputProps={{ min: 3, max: 6, step: 1 }}
-                                    type="number" />
-                            </div>
-                        </div>
-
-
-
-                        <div className={classes.settingsTitle}>
-                            <Typography>START INFO</Typography>
-                        </div>
-                        <div className={classes.onOffSwitchContainer}>
-                            <Typography variant="subtitle1" className={classes.onOffSwitchLabel}>Show start info</Typography>
-                            <div className={`${classes.onOffSwitch} off`}></div>
-                        </div>
-                    </div>
-
-                    <div className={classes.settingsRight}>
-                        <Typography>SET BALANCE & BLINDS</Typography>
-
-                        <div className={classes.setBalanceBlindsContainer}>
-                            <Typography className={classes.setBalanceBlindsTitle}>INITIAL BALANCE:</Typography>
-                            <div className={classes.balanceRow}>
-                                <Typography className={classes.inputLabel}>$</Typography>
-                                <TextField type="text" placeholder="0" className={classes.balanceInput} />
-                            </div>
-                            <div className={classes.blindRow}>
-                                <Typography className={classes.inputLabel}>BIG BLIND:</Typography>
-                                <TextField type="text" placeholder="0" className={classes.blindInput} />
-                                <Typography className={classes.inputLabel}>SMALL BLIND:</Typography>
-                                <TextField type="text" placeholder="0" className={classes.blindInput} />
-                            </div>
-                        </div>
-
-                        <Button variant="contained" color="primary" className={classes.startGameButton} onClick={handleStartGame}>START GAME</Button>
-                    </div>
-                </div>
+                                Start Game
+                            </Button>
+                            <Button variant={"contained"}
+                                    disabled={!isHost}
+                                    onClick={leaveGame}
+                            >
+                                Leave Game
+                            </Button>
+                            <Button className={classes.copyButton}
+                                    color={"inherit"}
+                                    variant={"outlined"}
+                                    onClick={() => {navigator.clipboard.writeText(gameId);}}
+                            >
+                                Copy lobby id
+                            </Button>
+                            <IconButton>
+                                <ExitToAppIcon/>
+                            </IconButton>
+                        </Toolbar>
+                    </AppBar>
+                    <Grid container className={classes.cardContainer}>
+                        <Grid item>
+                            {playerList}
+                        </Grid>
+                        <Grid item>
+                            {settings}
+                        </Grid>
+                    </Grid>
+                </Card>
             </Grid>
-        </div>
-    );
-
-    return lobby;
+        </Grid>
+    )
 }
 
 export default Lobby;
