+++ src/models/SettingsData.js
@@ -1,0 +1,14 @@
+/**
+ * Setting model
+ */
+class Settings {
+    constructor(data = {}) {
+        this.language = null;
+        this.playlistUrl = null;
+        this.balance = null;
+        this.bigBlind = null;
+        this.smallBlind = null;
+        Object.assign(this, data);
+    }
+}
+export default Settings;
